Update app-service-local-cache.md
By default, the local cache size is now set to 1 GB, with the option to increase it up to 2 GB (2,000 MB) per app.
One of my customers recently referred to this article and reached out to support to determine the currently allocated cache size. To make this article more helpful for future readers, we are updating it with this additional information.

By default, the local cache size is 1 GB. This size includes the /site and /siteextensions folders copied from the content store. It also includes any locally generated logs and data folders.

To increase this limit, use the app setting WEBSITE_LOCAL_CACHE_SIZEINMB. You can increase the size up to 2 GB (2,000 MB) per app. Keep in mind that a larger cache size increases the time to load the cache.

https://learn.microsoft.com/en-us/azure/app-service/overview-local-cache